% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{ct_geo_to}
\alias{ct_geo_to}
\title{Geocode data for 2011 census tracts.}
\format{A data frame with 65906 rows and 9 variables}
\source{
Statistics Canada
}
\description{
Geocoded census tract data filtered to the City of Toronto.
}
\details{
\itemize{
  \item long
  \item lat
  \item order
  \item hole
  \item piece
  \item group
  \item Geo_Code. The census tract ID
  \item ward. The city ward for the tract
  \item region. The region of the tract
  ...
}
}

