% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ct_positions.R
\name{ct_positions}
\alias{ct_positions}
\title{Estimate left-right score for each census tract.}
\usage{
ct_positions()
}
\value{
A dataframe containing the results from each worksheet.
}
\description{
Uses the candidate positions data to estimate left-right scores
}

