% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import_polling.R
\name{import_polling}
\alias{import_polling}
\title{Import polling data from an Excel file}
\usage{
import_polling()
}
\value{
A dataframe containing the polling breakdown.
}
\description{
Imports the polling data, broken down by agent characteristics
}

