% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/to_voters.R
\name{to_voters}
\alias{to_voters}
\title{Sample a specific number of Toronto voters.}
\usage{
to_voters(df, n = 100, export = FALSE)
}
\arguments{
\item{df}{The voters dataframe}

\item{n.}{The number of voters, defaults to 100}

\item{export.}{Should the voters be exported to csv}
}
\value{
A dataframe containing the voters.
}
\description{
Uses the toVoters dataset to sample a specific number of voters
}

