% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{toCensus}
\alias{toCensus}
\title{Data from the 2011 Census for Toronto.}
\format{A data frame with 2116555 rows and 8 variables}
\source{
Statistics Canada
}
\description{
Cleaned and filtered Census data for the Toronto CMA.
}
\details{
\itemize{
  \item Geo_Code. The census tract ID
  \item ward. The city ward for the tract
  \item region. The region of the tract
  \item age. Age categories
  \item gender. Male or female
  \item family_income. The median family income for the CT
  \item income_range. An ordered factor for income
  \item age_range. An ordered factor for age
  ...
}
}

